test(csharp): update tests for the RpcRegistry/MethodCaller split
Renames RpcBusDispatchTests -> MethodCallerDispatchTests,
RpcBusSerializationTests -> MethodCallerSerializationTests, and
RpcRegistrarAllTests -> RpcRegistryAllTests to track the production
rename, updating each test body to construct RpcRegistry + MethodCaller
directly instead of the old RpcBus/RpcRegistrar API.

Adds Tests/CSharpUtils/JsonRpcProtocol.cs: the 4-byte-framing JSON-RPC
parser removed from production code (now that StdioTransport implements
ITransportServer directly) survives here as a test-only injection path
for RpcIntegrationTests/UniLiquidLinkIntegrationTest, updated to the
little-endian frame length and RpcRequest-based parsing.

ConverterTests, JsonSerializerChainTests, ObjectRegistryTests,
RpcIntegrationTests, GenerateRpcGolden, and RpcTestContext follow the
RpcUnityObject -> RpcInstanceObject rename, the JsonSerializerChain
constructor taking no arguments (stage options now built internally
and exposed via chain.Options[(int)Stage]), and RpcResolveChainParam/
RpcResolveChainSetParam replacing positional chain-call arguments.
GoldenFiles/golden_resolve_chain_*.json and golden_find.json are
regenerated output reflecting those DTO shape changes.

+++ Tests/CSharp/ObjectRegistryTests.cs
@@ -41,8 +41,8 @@ public class ObjectRegistryTests
         var go = new GameObject("__ObjReg_Remove");
         try
         {
-            int id = (int)_reg.RegisterObject(go);
-            var removed = new List<int>();
+            long id = _reg.RegisterObject(go);
+            var removed = new List<long>();
             _reg.OnRemoveObject += i => removed.Add(i);
             _reg.RemoveObject(id);
             Assert.AreEqual(1, removed.Count);
